+++ examples/pencilbeam_helium.py
@@ -25,6 +25,7 @@ from pyRadPlan import (
     load_tg119,
 )
 
+from pyRadPlan.gui import launch_viewer, GUI_AVAILABLE
 
 logging.basicConfig(level=logging.INFO)
 
@@ -54,8 +55,12 @@ result = dij.compute_result_ct_grid(fluence)
 # %% [markdown]
 # Visualize the results
 # %%
-# Choose a slice to visualize
-view_slice = int(np.round(ct.size[2] / 2))
+if GUI_AVAILABLE:
+    # Use the GUI if [gui] dependencies are installed
+    launch_viewer(ct, cst, result)
+else:
+    # Choose a slice to visualize
+    view_slice = int(np.round(ct.size[2] / 2))
 
 # Visualize
 plot_slice(