+++ pyRadPlan/analysis/__init__.py
@@ -1,8 +1,18 @@
 """Treatment plan analysis tools and metrics."""
 
 from ._dvh import DVH, DVHCollection
+from ._qi import QI, QICollection, StructureQIs, Mean, Std, Max, Min, DX, VX
 
 __all__ = [
     "DVH",
     "DVHCollection",
+    "QI",
+    "QICollection",
+    "StructureQIs",
+    "Mean",
+    "Std",
+    "Max",
+    "Min",
+    "DX",
+    "VX",
 ]